+++ libs/preprocessor/linker.h
@@ -58,18 +58,9 @@ int lk_preprocess_all(environment *const env);
  *
  *	@param	env		Preprocessor environment
  *
- *	@return	@c  0 on success,
- *			@c -1 on file couldn't be opened, 
- *			@c -2 on error inside the file
+ *	@return	@c 0 on success, @c -1 on failure
  */
 int lk_include(environment *const env);
-
-/**
- *	Add a comment to indicate line changes in the output
- *
- *	@param	env	Preprocessor environment
- */
-void lk_add_comment(environment *const env);
 
 /**
  *	 Get current file name from linker
